refactor: write rewards log

+++ beacon-chain/core/altair/block.go
@@ -157,7 +157,7 @@ func ApplySyncRewardsPenalties(ctx context.Context, s state.BeaconStateAltair, v
 		if err != nil {
 			return nil, err
 		}
-		if err = helpers.LogBalanceChanges(uint64(index), balAtIdx, participantReward, newBalAtIdx, uint64(s.Slot()), nil, helpers.Increase, helpers.SyncCommittee); err != nil {
+		if err = helpers.LogBalanceChanges(index, balAtIdx, participantReward, newBalAtIdx, s.Slot(), nil, helpers.Increase, helpers.SyncCommittee); err != nil {
 			return nil, err
 		}
 		earnedProposerReward += proposerReward
@@ -183,7 +183,7 @@ func ApplySyncRewardsPenalties(ctx context.Context, s state.BeaconStateAltair, v
 	if err != nil {
 		return nil, err
 	}
-	if err = helpers.LogBalanceChanges(uint64(proposerIndex), balAtIdx, earnedProposerReward, newBalAtIdx, uint64(s.Slot()), nil, helpers.Increase, helpers.SyncProposer); err != nil {
+	if err = helpers.LogBalanceChanges(proposerIndex, balAtIdx, earnedProposerReward, newBalAtIdx, s.Slot(), nil, helpers.Increase, helpers.SyncProposer); err != nil {
 		return nil, err
 	}
 	// Apply sync committee penalties.
@@ -204,7 +204,7 @@ func ApplySyncRewardsPenalties(ctx context.Context, s state.BeaconStateAltair, v
 		if err != nil {
 			return nil, err
 		}
-		if err = helpers.LogBalanceChanges(uint64(index), balAtIdx, participantReward, newBalAtIdx, uint64(s.Slot()), nil, helpers.Decrease, helpers.SyncCommittee); err != nil {
+		if err = helpers.LogBalanceChanges(index, balAtIdx, participantReward, newBalAtIdx, s.Slot(), nil, helpers.Decrease, helpers.SyncCommittee); err != nil {
 			return nil, err
 		}
 	}
